Gazetteer provider implementation
