+++ web-ui/src/main/resources/catalog/locales/en-admin.json
@@ -1175,7 +1175,6 @@
     "oneLanguage": "One",
     "oneLanguage-help": "Choose one language to edit using the selector",
     "ui-mod-geocoder": "Gazetteer",
-    "ui-behaviour": "Behaviour (default)",
     "ui-geocoder-help": "Insert the url of Gazetteer service and the behaviour (alt or default)",
     "ui-mapExtent": "DEPRECATED! Map Extent",
     "ui-mapExtent-help": "DEPRECATED! Extent coordinates to use for maps other that viewer and search maps.",
